# apps/api/openapi_v4.json
{"openapi":"3.0.3","info":{"title":"7TV v4","version":"4.0.0","description":"Auto-generated v4 OpenAPI docs","license":{"name":"Apache-2.0 with Commons Clause"}},"paths":{"/v4/auth/login":{"get":{"summary":"Initiate login","responses":{"302":{"description":"Redirect to provider"}}}},"/v4/auth/link":{"get":{"summary":"Initiate account link","responses":{"302":{"description":"Redirect to provider"}}}},"/v4/auth/login/finish":{"post":{"summary":"Finish login","requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"description":"Login success"}}}},"/v4/auth/link/finish":{"post":{"summary":"Finish link","requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"description":"Link success"}}}},"/v4/auth/logout":{"post":{"summary":"Logout","responses":{"204":{"description":"Logged out"}}}},"/v4/gql":{"post":{"summary":"GraphQL endpoint","requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"description":"GraphQL response"}}}},"/v4/gql/playground":{"get":{"summary":"GraphQL playground","responses":{"200":{"description":"GraphQL playground HTML"}}}},"/v4/emotes/":{"post":{"summary":"Create emote","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object"}}}},"responses":{"201":{"description":"Created"}}}},"/v4/badges/":{"post":{"summary":"Create badge","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object"}}}},"responses":{"201":{"description":"Created"}}}},"/v4/events/create":{"post":{"summary":"Create event","requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"description":"Success"}}}},"/v4/users/{id}/profile-picture":{"post":{"summary":"Upload profile picture","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"image/*":{"schema":{"type":"string","format":"binary"}}}},"responses":{"200":{"description":"Uploaded"}}}},"/v4/users/{id}/products":{"get":{"summary":"Get user products","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}}}}